+++ src/app/(dashboard)/dashboard/providers/[id]/components/CustomModelsSection.tsx
@@ -1,0 +1,562 @@
+"use client";
+/**
+ * CustomModelsSection — Issue #3501 Phase 1e
+ *
+ * Extracted from ProviderDetailPageClient.tsx. Renders the "custom models"
+ * panel for ALL providers. This section is self-contained: it fetches its
+ * own model state from the API and manages local loading/saving state.
+ *
+ * Never imports from ProviderDetailPageClient.
+ */
+import React, { useState, useEffect, useCallback, useMemo } from "react";
+import { useTranslations } from "next-intl";
+import { Button } from "@/shared/components";
+import { useNotificationStore } from "@/store/notificationStore";
+import {
+  buildCompatMap,
+  anyNormalizeCompatBadge,
+  anyNoPreserveCompatBadge,
+  anyUpstreamHeadersBadge,
+  effectiveNormalizeForProtocol,
+  effectivePreserveForProtocol,
+  effectiveUpstreamHeadersForProtocol,
+  formatProviderModelsErrorResponse,
+  type CompatModelRow,
+  type CompatByProtocolMap,
+} from "../providerPageHelpers";
+import ModelCompatPopover from "./ModelCompatPopover";
+
+// ---------------------------------------------------------------------------
+// Props
+// ---------------------------------------------------------------------------
+
+export interface CustomModelsSectionProps {
+  providerId: string;
+  providerAlias: string;
+  copied?: string;
+  onCopy: (text: string, key: string) => void;
+  onModelsChanged?: () => void;
+}
+
+// ---------------------------------------------------------------------------
+// Component
+// ---------------------------------------------------------------------------
+
+export default function CustomModelsSection({
+  providerId,
+  providerAlias,
+  copied,
+  onCopy,
+  onModelsChanged,
+}: CustomModelsSectionProps) {
+  const t = useTranslations("providers");
+  const notify = useNotificationStore();
+  const [customModels, setCustomModels] = useState<CompatModelRow[]>([]);
+  const [modelCompatOverrides, setModelCompatOverrides] = useState<
+    Array<CompatModelRow & { id: string }>
+  >([]);
+  const [newModelId, setNewModelId] = useState("");
+  const [newModelName, setNewModelName] = useState("");
+  const [newApiFormat, setNewApiFormat] = useState("chat-completions");
+  const [newEndpoints, setNewEndpoints] = useState(["chat"]);
+  const [adding, setAdding] = useState(false);
+  const [loading, setLoading] = useState(true);
+  const [editingModelId, setEditingModelId] = useState<string | null>(null);
+  const [editingApiFormat, setEditingApiFormat] = useState("chat-completions");
+  const [editingEndpoints, setEditingEndpoints] = useState<string[]>(["chat"]);
+  const [savingModelId, setSavingModelId] = useState<string | null>(null);
+  const [togglingModelId, setTogglingModelId] = useState<string | null>(null);
+
+  const customMap = useMemo(() => buildCompatMap(customModels), [customModels]);
+  const overrideMap = useMemo(() => buildCompatMap(modelCompatOverrides), [modelCompatOverrides]);
+
+  const fetchCustomModels = useCallback(async () => {
+    try {
+      const res = await fetch(`/api/provider-models?provider=${encodeURIComponent(providerId)}`);
+      if (res.ok) {
+        const data = await res.json();
+        setCustomModels(data.models || []);
+        setModelCompatOverrides(data.modelCompatOverrides || []);
+      }
+    } catch (e) {
+      console.error("Failed to fetch custom models:", e);
+    } finally {
+      setLoading(false);
+    }
+  }, [providerId]);
+
+  useEffect(() => {
+    fetchCustomModels();
+  }, [fetchCustomModels]);
+
+  const handleAdd = async () => {
+    if (!newModelId.trim() || adding) return;
+    setAdding(true);
+    try {
+      const res = await fetch("/api/provider-models", {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({
+          provider: providerId,
+          modelId: newModelId.trim(),
+          modelName: newModelName.trim() || undefined,
+          apiFormat: newApiFormat,
+          supportedEndpoints: newEndpoints,
+        }),
+      });
+      if (res.ok) {
+        setNewModelId("");
+        setNewModelName("");
+        setNewApiFormat("chat-completions");
+        setNewEndpoints(["chat"]);
+        await fetchCustomModels();
+        onModelsChanged?.();
+      }
+    } catch (e) {
+      console.error("Failed to add custom model:", e);
+    } finally {
+      setAdding(false);
+    }
+  };
+
+  const handleRemove = async (modelId: string) => {
+    try {
+      await fetch(
+        `/api/provider-models?provider=${encodeURIComponent(providerId)}&model=${encodeURIComponent(modelId)}`,
+        {
+          method: "DELETE",
+        }
+      );
+      await fetchCustomModels();
+      onModelsChanged?.();
+    } catch (e) {
+      console.error("Failed to remove custom model:", e);
+    }
+  };
+
+  const handleToggleHidden = async (modelId: string, hidden: boolean) => {
+    setTogglingModelId(modelId);
+    try {
+      const res = await fetch(
+        `/api/provider-models?provider=${encodeURIComponent(providerId)}&modelId=${encodeURIComponent(modelId)}`,
+        {
+          method: "PATCH",
+          headers: { "Content-Type": "application/json" },
+          body: JSON.stringify({ isHidden: hidden }),
+        }
+      );
+      if (res.ok) {
+        await fetchCustomModels();
+        onModelsChanged?.();
+      }
+    } catch (e) {
+      console.error("Failed to toggle model visibility:", e);
+    } finally {
+      setTogglingModelId(null);
+    }
+  };
+
+  const beginEdit = (model: CompatModelRow) => {
+    setEditingModelId(model.id ?? null);
+    setEditingApiFormat(model.apiFormat || "chat-completions");
+    setEditingEndpoints(
+      Array.isArray(model.supportedEndpoints) && model.supportedEndpoints.length
+        ? model.supportedEndpoints
+        : ["chat"]
+    );
+  };
+
+  const cancelEdit = () => {
+    setEditingModelId(null);
+    setEditingApiFormat("chat-completions");
+    setEditingEndpoints(["chat"]);
+    setSavingModelId(null);
+  };
+
+  const saveCustomCompat = async (
+    modelId: string,
+    patch: { compatByProtocol?: CompatByProtocolMap }
+  ) => {
+    setSavingModelId(modelId);
+    try {
+      const res = await fetch("/api/provider-models", {
+        method: "PUT",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({ provider: providerId, modelId, ...patch }),
+      });
+      if (!res.ok) {
+        const detail = await formatProviderModelsErrorResponse(res);
+        notify.error(
+          detail ? `${t("failedSaveCustomModel")} — ${detail}` : t("failedSaveCustomModel")
+        );
+        return;
+      }
+    } catch {
+      notify.error(t("failedSaveCustomModel"));
+      return;
+    } finally {
+      setSavingModelId(null);
+    }
+    try {
+      await fetchCustomModels();
+      onModelsChanged?.();
+    } catch {
+      /* refresh failure is non-critical — data was already saved */
+    }
+  };
+
+  const saveEdit = async (modelId: string) => {
+    if (!editingModelId || editingModelId !== modelId) return;
+    if (!editingEndpoints.length) {
+      notify.error("Select at least one supported endpoint");
+      return;
+    }
+
+    setSavingModelId(modelId);
+    try {
+      const model = customModels.find((m) => m.id === modelId);
+      const res = await fetch("/api/provider-models", {
+        method: "PUT",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({
+          provider: providerId,
+          modelId,
+          modelName: model?.name || modelId,
+          source: model?.source || "manual",
+          apiFormat: editingApiFormat,
+          supportedEndpoints: editingEndpoints,
+        }),
+      });
+
+      if (!res.ok) {
+        const detail = await formatProviderModelsErrorResponse(res);
+        throw new Error(detail || "Failed to save model endpoint settings");
+      }
+
+      await fetchCustomModels();
+      onModelsChanged?.();
+      notify.success("Saved model endpoint settings");
+      cancelEdit();
+    } catch (e) {
+      console.error("Failed to save custom model:", e);
+      notify.error(
+        e instanceof Error && e.message ? e.message : "Failed to save model endpoint settings"
+      );
+    } finally {
+      setSavingModelId(null);
+    }
+  };
+
+  return (
+    <div className="mt-6 pt-6 border-t border-border">
+      <h3 className="text-sm font-semibold mb-3 flex items-center gap-2">
+        <span className="material-symbols-outlined text-base text-primary">tune</span>
+        {t("customModels")}
+      </h3>
+      <p className="text-xs text-text-muted mb-3">{t("customModelsHint")}</p>
+
+      {/* Add form */}
+      <div className="flex flex-col gap-3 mb-3">
+        <div className="flex items-end gap-2">
+          <div className="flex-1">
+            <label htmlFor="custom-model-id" className="text-xs text-text-muted mb-1 block">
+              {t("modelId")}
+            </label>
+            <input
+              id="custom-model-id"
+              type="text"
+              value={newModelId}
+              onChange={(e) => setNewModelId(e.target.value)}
+              onKeyDown={(e) => e.key === "Enter" && handleAdd()}
+              placeholder={t("customModelPlaceholder")}
+              className="w-full px-3 py-2 text-sm border border-border rounded-lg bg-background focus:outline-none focus:border-primary"
+            />
+          </div>
+          <div className="w-40">
+            <label htmlFor="custom-model-name" className="text-xs text-text-muted mb-1 block">
+              {t("displayName")}
+            </label>
+            <input
+              id="custom-model-name"
+              type="text"
+              value={newModelName}
+              onChange={(e) => setNewModelName(e.target.value)}
+              onKeyDown={(e) => e.key === "Enter" && handleAdd()}
+              placeholder={t("optional")}
+              className="w-full px-3 py-2 text-sm border border-border rounded-lg bg-background focus:outline-none focus:border-primary"
+            />
+          </div>
+          <Button size="sm" icon="add" onClick={handleAdd} disabled={!newModelId.trim() || adding}>
+            {adding ? t("adding") : t("add")}
+          </Button>
+        </div>
+
+        {/* API Format + Supported Endpoints */}
+        <div className="flex items-end gap-4 flex-wrap">
+          <div className="w-48">
+            <label htmlFor="custom-api-format" className="text-xs text-text-muted mb-1 block">
+              API Format
+            </label>
+            <select
+              id="custom-api-format"
+              value={newApiFormat}
+              onChange={(e) => setNewApiFormat(e.target.value)}
+              className="w-full px-3 py-2 text-sm border border-border rounded-lg bg-background focus:outline-none focus:border-primary"
+            >
+              <option value="chat-completions">{t("chatCompletions")}</option>
+              <option value="responses">{t("responsesApi")}</option>
+              <option value="embeddings">{t("embeddings")}</option>
+              <option value="rerank">Rerank</option>
+              <option value="audio-transcriptions">{t("audioTranscriptions")}</option>
+              <option value="audio-speech">{t("audioSpeech")}</option>
+              <option value="images-generations">{t("imagesGenerations")}</option>
+            </select>
+          </div>
+          <div className="flex-1">
+            <span className="text-xs text-text-muted mb-1 block">
+              {t("supportedEndpointsLabel")}
+            </span>
+            <div className="flex items-center gap-3">
+              {["chat", "embeddings", "rerank", "images", "audio"].map((ep) => (
+                <label
+                  key={ep}
+                  className="flex items-center gap-1.5 text-xs text-text-main cursor-pointer"
+                >
+                  <input
+                    type="checkbox"
+                    checked={newEndpoints.includes(ep)}
+                    onChange={(e) => {
+                      if (e.target.checked) {
+                        setNewEndpoints((prev) => [...prev, ep]);
+                      } else {
+                        setNewEndpoints((prev) => prev.filter((x) => x !== ep));
+                      }
+                    }}
+                    className="rounded border-border"
+                  />
+                  {ep === "chat"
+                    ? `💬 ${t("supportedEndpointChat")}`
+                    : ep === "embeddings"
+                      ? `📐 ${t("supportedEndpointEmbeddings")}`
+                      : ep === "rerank"
+                        ? "Rerank"
+                        : ep === "images"
+                          ? `🖼️ ${t("supportedEndpointImages")}`
+                          : `🔊 ${t("supportedEndpointAudio")}`}
+                </label>
+              ))}
+            </div>
+          </div>
+        </div>
+      </div>
+
+      {/* List */}
+      {loading ? (
+        <p className="text-xs text-text-muted">{t("loading")}</p>
+      ) : customModels.length > 0 ? (
+        <div className="flex flex-col gap-2">
+          {customModels.map((model) => {
+            const fullModel = `${providerAlias}/${model.id}`;
+            const copyKey = `custom-${model.id}`;
+            return (
+              <div
+                key={model.id}
+                className="flex items-center gap-3 rounded-lg border border-border p-3 hover:bg-sidebar/50"
+              >
+                {editingModelId !== model.id && (
+                  <span className="material-symbols-outlined text-base text-primary shrink-0">
+                    tune
+                  </span>
+                )}
+                <div className="min-w-0 flex-1">
+                  <p className="text-sm font-medium truncate">{model.name || model.id}</p>
+                  <div className="flex items-center gap-1 mt-1 flex-wrap">
+                    <code className="text-xs text-text-muted font-mono bg-sidebar px-1.5 py-0.5 rounded">
+                      {fullModel}
+                    </code>
+                    <button
+                      onClick={() => onCopy(fullModel, copyKey)}
+                      className="p-0.5 hover:bg-sidebar rounded text-text-muted hover:text-primary"
+                      title={t("copyModel")}
+                    >
+                      <span className="material-symbols-outlined text-sm">
+                        {copied === copyKey ? "check" : "content_copy"}
+                      </span>
+                    </button>
+                    {model.apiFormat === "responses" && (
+                      <span className="text-[10px] px-1.5 py-0.5 rounded-full bg-blue-500/15 text-blue-400 font-medium">
+                        {t("responses")}
+                      </span>
+                    )}
+                    {model.supportedEndpoints?.includes("embeddings") && (
+                      <span className="text-[10px] px-1.5 py-0.5 rounded-full bg-purple-500/15 text-purple-400 font-medium">
+                        {`📐 ${t("supportedEndpointEmbeddings")}`}
+                      </span>
+                    )}
+                    {model.supportedEndpoints?.includes("images") && (
+                      <span className="text-[10px] px-1.5 py-0.5 rounded-full bg-amber-500/15 text-amber-400 font-medium">
+                        {`🖼️ ${t("imagesShortLabel")}`}
+                      </span>
+                    )}
+                    {model.supportedEndpoints?.includes("audio") && (
+                      <span className="text-[10px] px-1.5 py-0.5 rounded-full bg-green-500/15 text-green-400 font-medium">
+                        {`🔊 ${t("audioShortLabel")}`}
+                      </span>
+                    )}
+                    {anyNormalizeCompatBadge(model.id!, customMap, overrideMap) && (
+                      <span
+                        className="text-[10px] px-1.5 py-0.5 rounded-full bg-slate-500/15 text-slate-400 font-medium"
+                        title={t("normalizeToolCallIdLabel")}
+                      >
+                        ID×9
+                      </span>
+                    )}
+                    {anyNoPreserveCompatBadge(model.id!, customMap, overrideMap) && (
+                      <span
+                        className="text-[10px] px-1.5 py-0.5 rounded-full bg-cyan-500/15 text-cyan-400 font-medium"
+                        title={t("compatDoNotPreserveDeveloper")}
+                      >
+                        {t("compatBadgeNoPreserve")}
+                      </span>
+                    )}
+                    {anyUpstreamHeadersBadge(model.id!, customMap, overrideMap) && (
+                      <span
+                        className="text-[10px] px-1.5 py-0.5 rounded-full bg-violet-500/15 text-violet-400 font-medium"
+                        title={t("compatUpstreamHeadersLabel")}
+                      >
+                        {t("compatBadgeUpstreamHeaders")}
+                      </span>
+                    )}
+                  </div>
+
+                  {editingModelId === model.id && (
+                    <div className="mt-3 min-w-0 max-w-full rounded-lg border border-border bg-muted p-3 dark:bg-zinc-900">
+                      <div className="flex min-w-0 flex-wrap items-end gap-x-3 gap-y-2">
+                        <div className="w-[11rem] shrink-0 min-w-0">
+                          <label className="text-xs text-text-muted mb-1 block">
+                            {t("apiFormatLabel")}
+                          </label>
+                          <select
+                            value={editingApiFormat}
+                            onChange={(e) => setEditingApiFormat(e.target.value)}
+                            className="w-full px-2.5 py-2 text-xs border border-border rounded-lg bg-background text-text-main focus:outline-none focus:border-primary"
+                          >
+                            <option value="chat-completions">{t("chatCompletions")}</option>
+                            <option value="responses">{t("responsesApi")}</option>
+                            <option value="embeddings">{t("embeddings")}</option>
+                            <option value="rerank">Rerank</option>
+                            <option value="audio-transcriptions">
+                              {t("audioTranscriptions")}
+                            </option>
+                            <option value="audio-speech">{t("audioSpeech")}</option>
+                            <option value="images-generations">{t("imagesGenerations")}</option>
+                          </select>
+                        </div>
+                        <div className="flex min-w-0 flex-1 flex-wrap items-center gap-x-3 gap-y-1 overflow-x-auto overflow-y-visible [scrollbar-width:thin]">
+                          <span className="text-xs text-text-muted shrink-0">
+                            {t("supportedEndpointsLabel")}
+                          </span>
+                          <div className="flex flex-wrap items-center gap-x-2 sm:gap-x-3 gap-y-1 min-w-0">
+                            {["chat", "embeddings", "rerank", "images", "audio"].map((ep) => (
+                              <label
+                                key={ep}
+                                className="flex items-center gap-1.5 text-xs text-text-main cursor-pointer whitespace-nowrap"
+                              >
+                                <input
+                                  type="checkbox"
+                                  checked={editingEndpoints.includes(ep)}
+                                  onChange={(e) => {
+                                    if (e.target.checked) {
+                                      setEditingEndpoints((prev) =>
+                                        prev.includes(ep) ? prev : [...prev, ep]
+                                      );
+                                    } else {
+                                      setEditingEndpoints((prev) => prev.filter((x) => x !== ep));
+                                    }
+                                  }}
+                                  className="rounded border-border"
+                                />
+                                {ep === "chat"
+                                  ? `💬 ${t("supportedEndpointChat")}`
+                                  : ep === "embeddings"
+                                    ? `📐 ${t("supportedEndpointEmbeddings")}`
+                                    : ep === "rerank"
+                                      ? "Rerank"
+                                      : ep === "images"
+                                        ? `🖼️ ${t("supportedEndpointImages")}`
+                                        : `🔊 ${t("supportedEndpointAudio")}`}
+                              </label>
+                            ))}
+                          </div>
+                        </div>
+                        <div className="flex shrink-0 flex-wrap items-center gap-2 pb-0.5">
+                          <Button
+                            size="sm"
+                            onClick={() => saveEdit(model.id!)}
+                            disabled={savingModelId === model.id}
+                          >
+                            {savingModelId === model.id ? t("saving") : t("save")}
+                          </Button>
+                          <Button size="sm" variant="ghost" onClick={cancelEdit}>
+                            {t("cancel")}
+                          </Button>
+                        </div>
+                      </div>
+                    </div>
+                  )}
+                </div>
+                <div className="flex shrink-0 items-center gap-1">
+                  <button
+                    onClick={() => beginEdit(model)}
+                    className="rounded p-1 text-text-muted hover:bg-sidebar hover:text-primary"
+                    title={t("edit")}
+                  >
+                    <span className="material-symbols-outlined text-sm">edit</span>
+                  </button>
+                  <ModelCompatPopover
+                    t={t}
+                    effectiveModelNormalize={(p) =>
+                      effectiveNormalizeForProtocol(model.id!, p, customMap, overrideMap)
+                    }
+                    effectiveModelPreserveDeveloper={(p) =>
+                      effectivePreserveForProtocol(model.id!, p, customMap, overrideMap)
+                    }
+                    getUpstreamHeadersRecord={(p) =>
+                      effectiveUpstreamHeadersForProtocol(model.id!, p, customMap, overrideMap)
+                    }
+                    onCompatPatch={(protocol, payload) =>
+                      saveCustomCompat(model.id!, {
+                        compatByProtocol: { [protocol]: payload },
+                      })
+                    }
+                    showDeveloperToggle
+                    disabled={savingModelId === model.id}
+                  />
+                  <button
+                    onClick={() => handleToggleHidden(model.id!, !model.isHidden)}
+                    disabled={togglingModelId === model.id}
+                    className="rounded p-1 text-text-muted hover:bg-sidebar hover:text-primary disabled:opacity-50"
+                    title={model.isHidden ? t("unhideModel") : t("hideModel")}
+                  >
+                    <span className="material-symbols-outlined text-sm">
+                      {model.isHidden ? "visibility_off" : "visibility"}
+                    </span>
+                  </button>
+                  <button
+                    onClick={() => handleRemove(model.id!)}
+                    className="rounded p-1 text-red-500 hover:bg-red-50"
+                    title={t("removeCustomModel")}
+                  >
+                    <span className="material-symbols-outlined text-sm">delete</span>
+                  </button>
+                </div>
+              </div>
+            );
+          })}
+        </div>
+      ) : (
+        <p className="text-xs text-text-muted">{t("noCustomModels")}</p>
+      )}
+    </div>
+  );
+}
